+++ tests/signature_pad.test.ts
@@ -1,7 +1,7 @@
 import SignaturePad from '../src/signature_pad';
 import { face } from './fixtures/face';
 import { square } from './fixtures/square';
-import './utils/pointer-event-polyfill'
+import './utils/pointer-event-polyfill';
 
 let canvas: HTMLCanvasElement;
 
@@ -119,22 +119,144 @@ describe('user interactions', () => {
         clientX: 50,
         clientY: 30,
         pressure: 1,
-      })
+      }),
     );
     canvas.dispatchEvent(
       new PointerEvent('pointerdown', {
         clientX: 240,
         clientY: 30,
         pressure: 1,
-      })
+      }),
     );
     canvas.dispatchEvent(
       new PointerEvent('pointerdown', {
         clientX: 150,
         clientY: 120,
         pressure: 1,
-      })
+      }),
     );
     expect(pad.toDataURL('image/svg+xml')).toMatchSnapshot();
+  });
+
+  it('call endStroke on pointerup outside canvas', () => {
+    const pad = new SignaturePad(canvas);
+    const endStroke = jest.fn();
+    pad.addEventListener('endStroke', endStroke);
+    canvas.dispatchEvent(
+      new PointerEvent('pointerdown', {
+        clientX: 50,
+        clientY: 30,
+        pressure: 1,
+      }),
+    );
+    canvas.dispatchEvent(
+      new PointerEvent('pointermove', {
+        clientX: 240,
+        clientY: 30,
+        pressure: 1,
+      }),
+    );
+    document.dispatchEvent(
+      new PointerEvent('pointerup', {
+        clientX: 150,
+        clientY: 120,
+        pressure: 1,
+      }),
+    );
+    expect(endStroke).toHaveBeenCalled();
+  });
+});
+
+describe('Signature events.', () => {
+  let signpad: SignaturePad;
+  let eventDispatched: Event | undefined;
+
+  const eventHandler: EventListener = (evt: Event): void => {
+    eventDispatched = evt;
+  };
+
+  beforeEach(() => {
+    signpad = new SignaturePad(canvas);
+
+    // to make this test works, canvas must be added to the document body.
+    document.body.insertAdjacentElement('afterbegin', canvas);
+
+    eventDispatched = undefined;
+  });
+
+  afterEach(() => {
+    document.body.removeChild(canvas);
+  });
+
+  [
+    { eventName: 'beginStroke', dispatchedEventName: ['pointerdown'] },
+    { eventName: 'beforeUpdateStroke', dispatchedEventName: ['pointerdown'] },
+    { eventName: 'afterUpdateStroke', dispatchedEventName: ['pointerdown'] },
+    {
+      eventName: 'endStroke',
+      dispatchedEventName: ['pointerdown', 'pointerup'],
+    },
+  ].forEach((param) => {
+    describe(`${param.eventName}.`, () => {
+      beforeEach(() => {
+        signpad.addEventListener(param.eventName, eventHandler);
+      });
+
+      afterEach(() => {
+        signpad.removeEventListener(param.eventName, eventHandler);
+      });
+
+      it('no writing to the canvas.', () => {
+        expect(eventDispatched).toBeFalsy();
+      });
+
+      it('writes to the canvas.', () => {
+        const eventInitObj = <PointerEventInit>{
+          clientX: 50,
+          clientY: 30,
+          pressure: 1,
+          bubbles: true,
+        };
+        let pointerEvent;
+        for (const dispatchedEventName of param.dispatchedEventName) {
+          pointerEvent = new PointerEvent(dispatchedEventName, eventInitObj);
+          canvas.dispatchEvent(pointerEvent);
+        }
+
+        expect(eventDispatched).toBeTruthy();
+        expect(eventDispatched).toBeInstanceOf(CustomEvent);
+
+        const event = <CustomEvent>eventDispatched;
+        expect(event.detail).toBe(pointerEvent);
+      });
+    });
+  });
+
+  describe(`use document as EventTarget.`, () => {
+    beforeEach(() => {
+      signpad['_et'] = document;
+
+      signpad.addEventListener('beginStroke', eventHandler);
+    });
+
+    afterEach(() => {
+      signpad.removeEventListener('beginStroke', eventHandler);
+    });
+
+    it('the event should be dispatched.', () => {
+      const eventInitObj = <PointerEventInit>{
+        clientX: 50,
+        clientY: 30,
+        pressure: 1,
+      };
+      const pointerEvent = new PointerEvent('pointerdown', eventInitObj);
+      canvas.dispatchEvent(pointerEvent);
+
+      expect(eventDispatched).toBeTruthy();
+      expect(eventDispatched).toBeInstanceOf(CustomEvent);
+
+      const event = <CustomEvent>eventDispatched;
+      expect(event.detail).toBe(pointerEvent);
+    });
   });
 });