+++ CHANGELOG.md
@@ -1,4 +1,23 @@
-## Changelog
+## [4.0.3](https://github.com/szimek/signature_pad/compare/v4.0.2...v4.0.3) (2022-03-18)
+
+
+### Bug Fixes
+
+* emit endStroke on pointerup outside of canvas ([#604](https://github.com/szimek/signature_pad/issues/604)) ([29b80dd](https://github.com/szimek/signature_pad/commit/29b80ddc7adcc923b4939e782a413ffc64ba3f5b))
+
+## [4.0.2](https://github.com/szimek/signature_pad/compare/v4.0.1...v4.0.2) (2022-01-21)
+
+
+### Bug Fixes
+
+* set user-select none on canvas ([#591](https://github.com/szimek/signature_pad/issues/591)) ([59ff331](https://github.com/szimek/signature_pad/commit/59ff3315a276ee3714c5dedce5ffc7014ba078ac))
+
+## [4.0.1](https://github.com/szimek/signature_pad/compare/v4.0.0...v4.0.1) (2022-01-08)
+
+
+### Bug Fixes
+
+* fix iOS <= 13 ([#581](https://github.com/szimek/signature_pad/pull/581))
 
 ### 4.0.0
 #### Bug fixes